pyproject.toml
[project]
name = "co-activation-grouping"
version = "0.1.0"
description = "Uses Delphi Co Activations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "evaluate>=0.4.5",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.5",
    "mlflow>=3.3.1",
    "mypy>=1.18.2",
    "networkx>=3.5",
    "pandas>=2.3.1",
    "pandas-stubs>=2.3.2.250926",
    "pydantic>=2.11.7",
    "pyvis>=0.3.2",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.1.0",
    "textstat>=0.7.8",
    "vllm==0.10.2",
    "eai-delphi"
]
[tool.uv]
package = true
[tool.uv.sources]
myproject = { path = "src" }
eai-delphi = { path = "../delphi" , editable = true }